requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib[bcrypt]
python-jose
beautifulsoup4
requests
python-multipart 
alembic   